LayoutTests/fast/css/webkit-line-clamp-calculated-value-expected.txt

Tests assigning a calculated value to -webkit-line-clamp CSS property.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS testDiv.style['-webkit-line-clamp'] is ""
testDiv.style['-webkit-line-clamp'] = 'calc(10% * 2)'
PASS testDiv.style['-webkit-line-clamp'] is "calc(20%)"
PASS window.getComputedStyle(testDiv).getPropertyValue('-webkit-line-clamp') is "20%"
testDiv.style['-webkit-line-clamp'] = 'calc(2 * 3)'
PASS testDiv.style['-webkit-line-clamp'] is "6"
PASS window.getComputedStyle(testDiv).getPropertyValue('-webkit-line-clamp') is "6"
PASS successfullyParsed is true

TEST COMPLETE
